requirements.txt
discord.py==1.3.2
mcstatus
jishaku
